M*SS: 1,修改id为long类型 2,修复bug

+++ app/src/main/java/com/xinyu/mwp/application/MyApplication.java
@@ -127,7 +127,7 @@ public class MyApplication extends MultiDexApplication implements OnUserUpdateLi
     @Override
     public void onUserUpdate(boolean isLogin) {
         String userToken = null;
-        int id = 0;
+        long id = 0;
         if (UserManager.getInstance().isLogin()) {
             userToken = UserManager.getInstance().getUserEntity().getToken();
             id = UserManager.getInstance().getUserEntity().getId();
@@ -253,5 +253,4 @@ public class MyApplication extends MultiDexApplication implements OnUserUpdateLi
             return -1;
         }
     }
-
 }
